Add ability to define resource_name at objects
