+++ tests/regime/submodule.py
@@ -4,17 +4,16 @@ Contains a class that is used to test the RegimeMeta class and hyperparameter de
 
 import torch
 
-from regime.utils import RegimeMeta, hyperparameter
+from regime import Node, hyperparameter
 
 
-class ExampleClassB(RegimeMeta):
+class ExampleClassB(Node):
     """
     Example class B that uses the RegimeMeta metaclass.
     """
 
-    @staticmethod
-    def resource_name() -> str:
-        return "example_b"
+    def __init__(self, resource_name: str):
+        super().__init__(resource_name)
 
     @hyperparameter("beta")
     def __call__(
